Move GT code to `gigl.nn`
