+++ gigl/nn/__init__.py
@@ -1,9 +1,17 @@
 """GiGL NN Module"""
 
+from gigl.nn.graph_transformer import (
+    FeedForwardNetwork,
+    GraphTransformerEncoder,
+    GraphTransformerEncoderLayer,
+)
 from gigl.nn.loss import RetrievalLoss
 from gigl.nn.models import LightGCN, LinkPredictionGNN
 
 __all__ = [
+    "FeedForwardNetwork",
+    "GraphTransformerEncoder",
+    "GraphTransformerEncoderLayer",
     "LightGCN",
     "LinkPredictionGNN",
     "RetrievalLoss",
